refactor(storage): introduce ILeaguesRepository and move GetLeaguesTeamPlaysIn from replays
