+++ src/Worms.Hub.Gateway/Ratings/RatingsCalculator.cs
@@ -7,6 +7,7 @@ using Worms.Hub.Storage.Domain;
 namespace Worms.Hub.Gateway.Ratings;
 
 internal sealed class RatingsCalculator(
+    ILeaguesRepository leaguesRepository,
     IReplaysRepository replaysRepository,
     ITeamsRepository teamsRepository,
     IRatingsRepository ratingsRepository,
@@ -80,7 +81,7 @@ internal sealed class RatingsCalculator(
         Justification = "ELO calculation failure for one league must not block remaining leagues or the claim/unclaim operation")]
     public void CalculateForTeam(string machine, string teamName)
     {
-        foreach (var leagueId in replaysRepository.GetLeaguesWithTeam(machine, teamName))
+        foreach (var leagueId in leaguesRepository.GetLeaguesTeamPlaysIn(machine, teamName))
         {
             try
             {
