+++ src/Worms.Hub.Storage/Database/ILeaguesRepository.cs
@@ -1,0 +1,11 @@
+using JetBrains.Annotations;
+
+namespace Worms.Hub.Storage.Database;
+
+[PublicAPI]
+public interface ILeaguesRepository
+{
+    IReadOnlyList<LeagueDb> GetAll();
+    LeagueDb? GetById(string id);
+    IReadOnlyList<string> GetLeaguesTeamPlaysIn(string machine, string teamName);
+}
