+++ src/Worms.Hub.Storage/Database/IReplaysRepository.cs
@@ -8,5 +8,4 @@ public interface IReplaysRepository
     Replay Create(Replay item);
     void Update(Replay item);
     IReadOnlyList<Replay> GetByLeagueId(string leagueId);
-    IReadOnlyList<string> GetLeaguesWithTeam(string machine, string teamName);
 }
