+++ src/Worms.Hub.Storage/Database/LeaguesRepository.cs
@@ -6,7 +6,7 @@ using Npgsql;
 namespace Worms.Hub.Storage.Database;
 
 [PublicAPI]
-public sealed class LeaguesRepository(IConfiguration configuration)
+public sealed class LeaguesRepository(IConfiguration configuration) : ILeaguesRepository
 {
     public IReadOnlyList<LeagueDb> GetAll()
     {
@@ -21,6 +21,20 @@ public sealed class LeaguesRepository(IConfiguration configuration)
         using var connection = new NpgsqlConnection(connectionString);
         return connection.QuerySingleOrDefault<LeagueDb>(
             "SELECT id, name FROM leagues WHERE id = @id", new { id });
+    }
+
+    public IReadOnlyList<string> GetLeaguesTeamPlaysIn(string machine, string teamName)
+    {
+        var connectionString = configuration.GetConnectionString("Database");
+        using var connection = new NpgsqlConnection(connectionString);
+        return [.. connection.Query<string>(
+            "SELECT DISTINCT r.league_id "
+            + "FROM replay_placements rp "
+            + "JOIN replays r ON r.id = rp.replay_id "
+            + "WHERE rp.machine = @machine AND rp.team_name = @teamName "
+            + "AND r.status = 'Processed' "
+            + "AND r.league_id IS NOT NULL",
+            new { machine, teamName })];
     }
 }
